requirements.txt
absl-py==0.9.0
afinn==0.1
algorithm==1.2.0
ansi2html==1.5.2
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.5.3
billiard==3.6.3.0
biopython==1.68
bleach==3.1.5
blis==0.4.1
Brotli==1.0.9
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cmdstanpy==0.4.0
coala-utils==0.5.1
common==0.1.2
community==1.0.0b1
convertdate==2.2.1
cryptography==1.7.1
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.25.2
dash==1.16.2
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
datasketch==1.5.1
decorator==4.4.2
defusedxml==0.6.0
dependency-injection==1.2.0
distlib==0.3.1
elasticsearch==7.1.0
entrypoints==0.3
ephem==3.7.7.1
extras==0.0.3
fbprophet==0.6
filelock==3.0.12
findspark==1.4.2
Fiona==1.8.17
Flask==1.1.1
Flask-Compress==1.5.0
future==0.18.2
gast==0.3.3
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
GetOldTweets3==0.0.11
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googlemaps==4.4.2
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
holidays==0.10.3
html5lib==0.999999999
hupper==1.10.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.2.1
install==1.3.4
interruptingcow==0.8
iotop==0.6
ipykernel==5.3.4
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.1
joblib==0.10.4.dev0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyter-dash==0.3.1
kafka==1.3.5
kazam==1.4.5
kazoo==2.5.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==10.1
keyrings.alt==1.3
korean-lunar-calendar==0.2.1
linecache2==1.0.0
louis==3.0.0
LunarCalendar==0.0.9
lxml==3.7.1
Mako==1.0.6
Markdown==3.2.2
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.8.4
mpi4py==2.0.0
munch==2.5.0
murmurhash==1.0.2
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
nose==1.3.7
nose2==0.6.5
notebook==6.0.3
numexpr==2.6.1
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.1
PasteDeploy==2.1.0
patsy==0.5.1
pbr==1.10.0
pep8==1.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==4.0.0
plac==1.1.3
plaster==1.0
plaster-pastedeploy==0.7
plotly==4.5.4
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==2.0.10
prophet==0.1.1.post1
protobuf==3.13.0
ptyprocess==0.6.0
py==1.4.32
py4j==0.10.9
pyasn1==0.1.9
pyasn1-modules==0.2.8
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.3.0
Pygments==2.2.0
pygobject==3.22.0
pykafka==2.8.0
PyMeeus==0.3.7
pymongo==3.10.1
pyparsing==2.1.10
pyproj==2.6.1.post1
pyquery==1.4.1
pyramid-arima==0.9.0
pyrsistent==0.16.0
PySocks==1.7.1
pyspark==3.0.0
pystan==2.19.1.1
pytest==3.0.6
python-apt==1.4.1
python-dateutil==2.8.1
python-debian==0.1.30
python-igraph==0.7.1.post6
python-mimeparse==0.1.4
python2==1.2
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
regex==2020.7.14
reportlab==3.3.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rpy2==2.8.5
rsa==4.6
scikit-image==0.12.3
scikit-learn==0.18
scipy==1.4.1
seaborn==0.9.1
SecretStorage==2.3.1
Send2Trash==1.5.0
setuptools-git==1.2
Shapely==1.7.1
simplejson==3.10.0
six==1.15.0
sklearn==0.0
spacy==2.3.2
spark-nlp==2.5.4
SQLAlchemy==1.3.12
srsly==1.0.2
statsmodels==0.11.1
tables==3.3.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testtools==1.8.0
textblob==0.15.3
thinc==7.4.1
tornado==6.0.4
tqdm==4.48.2
traceback2==1.4.0
traitlets==4.3.3
translationstring==1.4
tweepy==3.9.0
twitterscraper==1.6.1
unattended-upgrades==0.1
Unidecode==1.1.1
unittest2==1.1.0
urllib3==1.25.7
usb-creator==0.3.0
vaderSentiment==3.3.2
venusian==3.0.0
virtualenv==20.1.0
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5
WebOb==1.8.6
Werkzeug==1.0.0
wget==3.2
wordcloud==1.8.0
wrapt==1.12.1
xgboost==1.1.1
youtube-dl==2017.5.18.1
zipp==1.2.0
zope.deprecation==4.4.0
zope.interface==5.1.0
